Feature/bug fix round one

+++ AscendNashville/src/main/java/com/nashss/se/ascendnashville/activity/DeleteEventActivity.java
@@ -1,12 +1,15 @@
 package com.nashss.se.ascendnashville.activity;
 
 
+import com.nashss.se.ascendnashville.Exceptions.EventNotFoundException;
+import com.nashss.se.ascendnashville.Exceptions.InvalidAttributeValueException;
 import com.nashss.se.ascendnashville.activity.requests.DeleteEventRequest;
 import com.nashss.se.ascendnashville.activity.results.DeleteEventResult;
 
 import com.nashss.se.ascendnashville.converters.ModelConverter;
 import com.nashss.se.ascendnashville.dynamoDB.EventDao;
 import com.nashss.se.ascendnashville.dynamoDB.models.Event;
+import com.nashss.se.ascendnashville.utils.AscendNashvilleUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -41,16 +44,22 @@ public class DeleteEventActivity {
      * If the provided eventId has invalid characters, throws an
      * InvalidAttributeValueException
      * <p>
-     * If the request tries to update the event Id,
-     * this should throw an InvalidAttributeChangeException
-     *
      * @param deleteEventRequest request object containing the event ID
      *                              associated with it
      */
     public DeleteEventResult handleRequest(final DeleteEventRequest deleteEventRequest) {
         log.info("Inside DeleteEventActivity handleRequest.");
+
+        if(!AscendNashvilleUtils.isValidString(deleteEventRequest.getEventId())) {
+            throw new InvalidAttributeValueException("Event date [" + deleteEventRequest.getEventId() +
+                    "} contains illegal characters");
+        }
         Event event = eventDao.getEvent(deleteEventRequest.getEventId());
 
+        if (event == null) {
+            throw new EventNotFoundException("No event exists associated with event Id:"
+                    + deleteEventRequest.getEventId());
+        }
         event.setEventId(deleteEventRequest.getEventId());
         eventDao.deleteEvent(event);
